+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Feat
 
+- **pic**: bump PocketIC to v14, add disableIngressValidation and senderInfo options
 - **pic**: add costSchedule option and bump PocketIC to v13 (#264)
 
 ## 0.21.0 (2026-03-18)